requirements.txt
anyio==4.2.0
appdirs==1.4.4
asttokens==2.4.1
beautifulsoup4==4.12.3
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
dhanhq==1.2.3
executing==2.0.1
fonttools==4.47.2
frozendict==2.4.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.2
httpx==0.25.2
idna==3.6
ipykernel==6.29.0
ipython==8.20.0
jedi==0.19.1
jupyter_client==8.6.0
jupyter_core==5.7.1
kiwisolver==1.4.5
lxml==5.1.0
matplotlib==3.8.2
matplotlib-inline==0.1.6
multitasking==0.0.11
nest-asyncio==1.5.9
numpy==1.26.3
packaging==23.2
pandas==2.1.4
parso==0.8.3
peewee==3.17.0
pillow==10.2.0
platformdirs==4.1.0
prompt-toolkit==3.0.43
psutil==5.9.7
pure-eval==0.2.2
Pygments==2.17.2
pyparsing==3.1.1
pyTelegramBotAPI==4.15.2
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
pyzmq==25.1.2
requests==2.31.0
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
tornado==6.4
traitlets==5.14.1
tzdata==2023.4
urllib3==2.1.0
wcwidth==0.2.13
webencodings==0.5.1
yfinance==0.2.35
